+++ src/components/layout/nav.js
@@ -7,7 +7,7 @@ import { LocalizedLink, localize } from 'components/localization'
 import { Button } from 'components/form'
 import { Container } from 'components/containers'
 import { Modal, useModal, OffCanvasMenu, moveOffCanvasMenu } from 'components/elements'
-import SignupModal from 'components/custom/signup-modal2'
+import SignupModal from 'components/custom/signup-modal'
 import { SharedLinkStyle } from 'components/localization/localized-link'
 import Login from 'common/login'
 import device from 'themes/device'
